Misc. cleanups
